new UseMinimalResources parameter + replaced bits transfer with System.Net.WebClient

+++ bootstrap/primary-domain-controller-step-1.ps1
@@ -41,6 +41,19 @@ Function Unwrap-SecureString() {
 	Return (New-Object -TypeName System.Net.NetworkCredential -ArgumentList '', $SecureString).Password
 }
 
+Function Get-GoogleMetadata() {
+        Param (
+        [Parameter(Mandatory=$True)][String] $Path
+        )
+        Try {
+                Return Invoke-RestMethod -Headers @{"Metadata-Flavor" = "Google"} -Uri http://169.254.169.254/computeMetadata/v1/$Path
+        }
+        Catch {
+                Return $Null
+        }
+}
+
+
 Write-Host "Bootstrap script started..."
 
 
@@ -57,11 +70,13 @@ Install-WindowsFeature -name AD-Domain-Services -IncludeManagementTools
 #	$zone = Invoke-RestMethod -Headers @{"Metadata-Flavor" = "Google"} -Uri http://169.254.169.254/computeMetadata/v1/instance/zone
 #	gcloud compute instances delete-access-config $name --zone $zone
 #}
-Write-Host "Removing external address..."
-# windows should have activated before script is invoked, so now remove external address
-$name = Invoke-RestMethod -Headers @{"Metadata-Flavor" = "Google"} -Uri http://169.254.169.254/computeMetadata/v1/instance/name
-$zone = Invoke-RestMethod -Headers @{"Metadata-Flavor" = "Google"} -Uri http://169.254.169.254/computeMetadata/v1/instance/zone
-gcloud compute instances delete-access-config $name --zone $zone
+
+If ("true" -like (Get-GoogleMetadata "instance/attributes/remove-address")) {
+        Write-Host "Removing external address..."
+        $name = Get-GoogleMetadata "instance/name"
+        $zone = Get-GoogleMetadata "instance/zone"
+        gcloud compute instances delete-access-config $name --zone $zone
+}
 
 
 Write-Host "Configuring network..."
